+++ src/leapfrogai_api/main.py
@@ -18,7 +18,6 @@ from leapfrogai_api.routers.openai import (
     threads,
     messages,
     runs,
-    runs_steps,
     vector_stores,
 )
 from leapfrogai_api.utils import get_model_config
@@ -64,8 +63,7 @@ app.include_router(embeddings.router)
 app.include_router(assistants.router)
 app.include_router(files.router)
 app.include_router(vector_stores.router)
-app.include_router(runs.router)
 app.include_router(threads.router)
 app.include_router(messages.router)
-app.include_router(runs_steps.router)
+app.include_router(runs.router)
 app.include_router(rag.router)
